Fix indexer when behavior is not activated.
Do not let themes and categories end up in the catalog index when the
behavior is not activated for a portal_type.

Say you have a container type with the behaviors activated, and an
item type that does not have them activated.  You give a container
instance the theme 'Water'.  You add an item in the container.  The
item gets indexed.  Until now, our indexer was only registered for
types with the themes behavior enabled, so for items Plone falls back
on the standard indexer that simply gets the property of that name.
This property is fetched from the parent due to acquisition...

Now we register our indexers for Interface, so for all types, and make
it return None if the relevant behavior is not activated.